# backend/api-collections/fleet-manager/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/drones": {
      "get": {
        "operationId": "DroneController_getDronesAsync",
        "summary": "Lists drones",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "model",
            "required": false,
            "in": "query",
            "description": "Filter by drone model",
            "schema": {
              "$ref": "#/components/schemas/DroneModel"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by drone status",
            "schema": {
              "$ref": "#/components/schemas/DroneStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of drones has been successfully retrieved."
          }
        },
        "tags": [
          "drones"
        ]
      },
      "post": {
        "operationId": "DroneController_createDroneAsync",
        "summary": "Creates a new drone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data for the drone to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDroneRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The drone has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDroneResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing data."
          }
        },
        "tags": [
          "drones"
        ]
      }
    },
    "/drones/{id}": {
      "get": {
        "operationId": "DroneController_getDroneAsync",
        "summary": "Retrieves information about a specific drone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the drone",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The drone information has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDroneResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Drone not found."
          }
        },
        "tags": [
          "drones"
        ]
      },
      "put": {
        "operationId": "DroneController_updateDroneAsync",
        "summary": "Updates an existing drone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the drone to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The data for the drone to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDroneRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The drone has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDroneResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing data."
          },
          "404": {
            "description": "Drone not found."
          }
        },
        "tags": [
          "drones"
        ]
      },
      "delete": {
        "operationId": "DroneController_softDeleteDroneAsync",
        "summary": "Deletes a specific drone (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the drone to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The drone has been successfully deleted."
          },
          "404": {
            "description": "Drone not found."
          }
        },
        "tags": [
          "drones"
        ]
      }
    },
...
    "/maintenance": {
      "get": {
        "operationId": "MaintenanceController_getMaintenanceLogsAsync",
        "summary": "Lists maintenance logs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "droneId",
            "required": false,
            "in": "query",
            "description": "Filter by drone ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by maintenance type",
            "schema": {
              "$ref": "#/components/schemas/MaintenanceType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of maintenance logs has been successfully retrieved."
          }
        },
        "tags": [
          "maintenance"
        ]
      },
      "post": {
        "operationId": "MaintenanceController_createMaintenanceLogAsync",
        "summary": "Creates a new maintenance log",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data for the maintenance log to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceLogRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The maintenance log has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMaintenanceLogResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing data."
          }
        },
        "tags": [
          "maintenance"
        ]
      }
    },
    "/maintenance/{id}": {
      "get": {
        "operationId": "MaintenanceController_getMaintenanceLogAsync",
        "summary": "Retrieves information about a specific maintenance log",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the maintenance log",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The maintenance log information has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMaintenanceLogResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance log not found."
          }
        },
        "tags": [
          "maintenance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet Manager API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
      "DroneModel": {
        "type": "string",
        "description": "Model of the drone",
        "enum": [
          "PHANTOM_4",
          "MATRICE_300",
          "MAVIC_3_ENTERPRISE"
        ]
      },
      "DroneStatus": {
        "type": "string",
        "description": "Status of the drone (Manual update is restricted to RETIRED only)",
        "enum": [
          "RETIRED"
        ]
      },
      "GetDroneResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the drone"
          },
          "serialNumber": {
            "type": "string",
            "description": "Drone unique serial number (Format: SKY-XXXX-XXXX)"
          },
          "model": {
            "type": "string",
            "description": "Drone model",
            "enum": [
              "PHANTOM_4",
              "MATRICE_300",
              "MAVIC_3_ENTERPRISE"
            ]
          },
          "status": {
            "type": "string",
            "description": "Current status of the drone",
            "enum": [
              "AVAILABLE",
              "IN_MISSION",
              "MAINTENANCE",
              "RETIRED"
            ]
          },
          "totalFlightHours": {
            "type": "number",
            "description": "Total flight hours accumulated"
          },
          "flightHoursAtLastMaintenance": {
            "type": "number",
            "description": "Flight hours at last maintenance"
          },
          "lastMaintenanceDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date of the last maintenance"
          },
          "nextMaintenanceDueDate": {
            "format": "date-time",
            "type": "string",
            "description": "Due date for the next maintenance"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record update timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "model",
          "status",
          "totalFlightHours",
          "flightHoursAtLastMaintenance",
          "lastMaintenanceDate",
          "nextMaintenanceDueDate",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "CreateDroneRequestDto": {
        "type": "object",
        "properties": {
          "serialNumber": {
            "type": "string",
            "description": "Serial number of the drone (exactly 13 characters long)",
            "example": "SKY-AB12-CD34"
          },
          "model": {
            "example": "MATRICE_300",
            "$ref": "#/components/schemas/DroneModel"
          }
        },
        "required": [
          "serialNumber",
          "model"
        ]
      },
      "CreateDroneResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the drone"
          },
          "serialNumber": {
            "type": "string",
            "description": "Drone unique serial number (Format: SKY-XXXX-XXXX)"
          },
          "model": {
            "type": "string",
            "description": "Drone model",
            "enum": [
              "PHANTOM_4",
              "MATRICE_300",
              "MAVIC_3_ENTERPRISE"
            ]
          },
          "status": {
            "type": "string",
            "description": "Current status of the drone",
            "enum": [
              "AVAILABLE",
              "IN_MISSION",
              "MAINTENANCE",
              "RETIRED"
            ]
          },
          "totalFlightHours": {
            "type": "number",
            "description": "Total flight hours accumulated"
          },
          "flightHoursAtLastMaintenance": {
            "type": "number",
            "description": "Flight hours at last maintenance"
          },
          "lastMaintenanceDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date of the last maintenance"
          },
          "nextMaintenanceDueDate": {
            "format": "date-time",
            "type": "string",
            "description": "Due date for the next maintenance"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "model",
          "status",
          "totalFlightHours",
          "flightHoursAtLastMaintenance",
          "lastMaintenanceDate",
          "nextMaintenanceDueDate",
          "createdAt",
          "version"
        ]
      },
      "UpdateDroneRequestDto": {
        "type": "object",
        "properties": {
          "model": {
            "example": "MATRICE_300",
            "$ref": "#/components/schemas/DroneModel"
          },
          "status": {
            "example": "RETIRED",
            "$ref": "#/components/schemas/DroneStatus"
          },
          "version": {
            "type": "number",
            "description": "Current row version for optimistic locking",
            "example": 1
          }
        },
        "required": [
          "version"
        ]
      },
      "UpdateDroneResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the drone"
          },
          "serialNumber": {
            "type": "string",
            "description": "Drone unique serial number (Format: SKY-XXXX-XXXX)"
          },
          "model": {
            "type": "string",
            "description": "Drone model",
            "enum": [
              "PHANTOM_4",
              "MATRICE_300",
              "MAVIC_3_ENTERPRISE"
            ]
          },
          "status": {
            "type": "string",
            "description": "Current status of the drone",
            "enum": [
              "AVAILABLE",
              "IN_MISSION",
              "MAINTENANCE",
              "RETIRED"
            ]
          },
          "totalFlightHours": {
            "type": "number",
            "description": "Total flight hours accumulated"
          },
          "flightHoursAtLastMaintenance": {
            "type": "number",
            "description": "Flight hours at last maintenance"
          },
          "lastMaintenanceDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date of the last maintenance"
          },
          "nextMaintenanceDueDate": {
            "format": "date-time",
            "type": "string",
            "description": "Due date for the next maintenance"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record update timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "model",
          "status",
          "totalFlightHours",
          "flightHoursAtLastMaintenance",
          "lastMaintenanceDate",
          "nextMaintenanceDueDate",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
      "MaintenanceType": {
        "type": "string",
        "description": "Type of maintenance performed",
        "enum": [
          "ROUTINE_CHECK",
          "BATTERY_REPLACEMENT",
          "MOTOR_REPAIR",
          "FIRMWARE_UPDATE",
          "FULL_OVERHAUL"
        ]
      },
      "GetMaintenanceLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the maintenance log"
          },
          "droneId": {
            "type": "string",
            "description": "ID of the drone"
          },
          "type": {
            "type": "string",
            "description": "Type of maintenance performed",
            "enum": [
              "ROUTINE_CHECK",
              "BATTERY_REPLACEMENT",
              "MOTOR_REPAIR",
              "FIRMWARE_UPDATE",
              "FULL_OVERHAUL"
            ]
          },
          "technicianName": {
            "type": "string",
            "description": "Technician name"
          },
          "notes": {
            "type": "string",
            "description": "Maintenance notes"
          },
          "performedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date maintenance was performed"
          },
          "flightHoursAtMaintenance": {
            "type": "number",
            "description": "Flight hours recorded at maintenance"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record update timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "droneId",
          "type",
          "technicianName",
          "notes",
          "performedAt",
          "flightHoursAtMaintenance",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "CreateMaintenanceLogRequestDto": {
        "type": "object",
        "properties": {
          "droneId": {
            "type": "string",
            "description": "ID of the drone being maintained",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "type": {
            "example": "ROUTINE_CHECK",
            "$ref": "#/components/schemas/MaintenanceType"
          },
          "technicianName": {
            "type": "string",
            "description": "Name of the technician performing maintenance",
            "example": "John Doe"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes or remarks",
            "example": "Replaced left rotor and updated firmware."
          },
          "performedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when maintenance was performed",
            "example": "2026-07-26T10:00:00.000Z"
          },
          "flightHoursAtMaintenance": {
            "type": "number",
            "description": "Drone flight hours at the time of maintenance",
            "example": 45.5
          }
        },
        "required": [
          "droneId",
          "type",
          "technicianName",
          "performedAt",
          "flightHoursAtMaintenance"
        ]
      },
      "CreateMaintenanceLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the maintenance log"
          },
          "droneId": {
            "type": "string",
            "description": "ID of the drone"
          },
          "type": {
            "type": "string",
            "description": "Type of maintenance performed",
            "enum": [
              "ROUTINE_CHECK",
              "BATTERY_REPLACEMENT",
              "MOTOR_REPAIR",
              "FIRMWARE_UPDATE",
              "FULL_OVERHAUL"
            ]
          },
          "technicianName": {
            "type": "string",
            "description": "Technician name"
          },
          "notes": {
            "type": "string",
            "description": "Maintenance notes"
          },
          "performedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date maintenance was performed"
          },
          "flightHoursAtMaintenance": {
            "type": "number",
            "description": "Flight hours recorded at maintenance"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "droneId",
          "type",
          "technicianName",
          "notes",
          "performedAt",
          "flightHoursAtMaintenance",
          "createdAt",
          "version"
        ]
      },
...
    }
  }
}